+++ crates/testing/src/launch.rs
@@ -42,7 +42,7 @@ pub fn launch_test_with_config(root: AppComponent, config: TestingConfig) -> Tes
         preferred_theme: PreferredTheme::default(),
         navigation_mode: NavigationMode::default(),
         information: PlatformInformation::new(config.size, false, false, false),
-        scale_factor: SCALE_FACTOR as f32,
+        scale_factor: SCALE_FACTOR,
     });
     let mut font_collection = FontCollection::new();
     let font_mgr = FontMgr::default();
